requirements.txt
torch
torchvision
datasets
transformers
accelerate
bitsandbytes
sentencepiece
protobuf
pymupdf
langchain
langchain-community
pdfplumber
numpy
sentence-transformers
faiss_cpu
ollama
openai
scikit-learn
fastapi
typing
pydantic
python-multipart
